REFACTOR: API shaping: - splitted apart contract transaction controller for specific types; - clarified naming of portfolio controllers; - implemented sub/childcontrollers logic for overview; - contract violations makes sense under singe contract, so put it's acces under single contract.
