+++ src/controllers/api/ContractSubControllers.ts
@@ -1,0 +1,14 @@
+import { Controller } from "@tsed/di";
+import { ContractTransactionsCreditController } from "./contracts/ContractTransactionsCreditController";
+import { ContractTransactionsDebitController } from "./contracts/ContractTransactionsDebitController";
+import { ContractViolationsController } from "./contracts/ContractViolationsController";
+
+@Controller({
+  path: "/contracts/:contractId",
+  children: [
+    ContractViolationsController,
+    ContractTransactionsDebitController,
+    ContractTransactionsCreditController
+  ]
+})
+export class ContractSubControllers {}
