--- src/controllers/api/TransactionsController.ts
@@ -1,9 +1,0 @@
-import { Controller } from "@tsed/di";
-import { TransactionsPortfolioCreditController } from "./transactions/TransactionsPortfolioCreditController";
-import { TransactionsPortfolioDebitController } from "./transactions/TransactionsPortfolioDebitController";
-
-@Controller({
-  path: "/transactions",
-  children: [TransactionsPortfolioCreditController, TransactionsPortfolioDebitController]
-})
-export class TransactionsController {}
