+++ src/controllers/api/contracts/ContractViolationsController.ts
@@ -1,8 +1,8 @@
 import { RolesEnum } from "@/common/RolesEnum";
 import { ShapesEnum } from "@/common/ShapesEnum";
 import { AuthorizedRoles } from "@/middlewares/AuthorizedRoles";
-import { DebitPortfolioTransactionModel } from "@/models/DebitPortfolioTransactionModel";
-import { DebitPortfolioTransactionService } from "@/services/DebitPortfolioTransactionService";
+import { ContractViolationModel } from "@/models/ContractViolationModel";
+import { ContractViolationService } from "@/services/ContractViolationService";
 import { OASDocs } from "@/utils/OASDocs";
 import { Controller, Inject } from "@tsed/di";
 import { Authenticate } from "@tsed/passport";
@@ -21,31 +21,31 @@ import {
 } from "@tsed/schema";
 import { BaseController } from "../common/BaseController";
 
-export const d = new OASDocs("portfolio debit transaction");
+const d = new OASDocs("contract violation");
 
-@Controller("/portfolio/debit")
+@Controller("/violations")
 @Description(d.getControllerDecr())
 @Security("jwt")
 @Authenticate("jwt", { session: false })
 @AuthorizedRoles(RolesEnum.LENDER)
 @Status(400).Description(OASDocs.STATUS_400_DESCR_VALIDATION)
 @Status(401).Description(OASDocs.STATUS_401_DESCR)
-export class TransactionsPortfolioDebitController extends BaseController {
-  constructor(@Inject() private service: DebitPortfolioTransactionService) {
+export class ContractViolationsController extends BaseController {
+  constructor(@Inject() private service: ContractViolationService) {
     super();
   }
 
   // TODO: https://tsed.io/docs/model.html#pagination
   @Get()
   @Summary(d.getAllSummary())
-  @Status(200, Array).Of(DebitPortfolioTransactionModel)
+  @Status(200, Array).Of(ContractViolationModel)
   async get() {
     return await this.service.getAll();
   }
 
   @Get("/:id")
   @Summary(d.getDocId())
-  @Status(200, DebitPortfolioTransactionModel)
+  @Status(200, ContractViolationModel)
   @Status(404).Description(d.getNoDoc())
   async getId(@Description(d.getGetParamId()) @PathParams() @Required() { id }: never) {
     const objModel = await this.service.findById(id);
@@ -57,25 +57,25 @@ export class TransactionsPortfolioDebitController extends BaseController {
 
   @Post()
   @Summary(d.getPostSummary())
-  @Status(201, DebitPortfolioTransactionModel).Description(d.getPost201StatusDescr())
+  @Status(201, ContractViolationModel).Description(d.getPost201StatusDescr())
   async post(
     @Description(d.getParamPostDtoDescr())
     @BodyParams()
     @Required()
     @Groups(ShapesEnum.CRE)
-    dto: DebitPortfolioTransactionModel
+    dto: ContractViolationModel
   ) {
     return await this.service.create(dto);
   }
 
   @Put("/:id")
-  @Summary("Update an account")
+  @Summary(d.getPutSummary())
   @Status(204).Description("Updated")
   @Status(400).Description(OASDocs.STATUS_400_ID_MISMATCH)
   @Status(404).Description(d.get404ForNonExisting("update"))
   async put(
-    @Description(d.getParamPutIdDescr()) @PathParams() @Required() { id }: never,
-    @Description(d.getParamPutDtoDescr()) @BodyParams() dto: DebitPortfolioTransactionModel
+    @Description(d.getParamPutIdDescr()) @Required() { id }: never,
+    @Description(d.getParamPutDtoDescr()) @BodyParams() dto: ContractViolationModel
   ) {
     BaseController.assertPutFixIfPossible(id, dto);
