+++ src/controllers/api/portfolio/PortfolioTransactionsCreditController.ts
@@ -23,14 +23,14 @@ import { BaseController } from "../common/BaseController";
 
 const d = new OASDocs("portfolio credit transaction");
 
-@Controller("/portfolio/credit")
+@Controller("/transactions/credit")
 @Description(d.getControllerDecr())
 @Security("jwt")
 @Authenticate("jwt", { session: false })
 @AuthorizedRoles(RolesEnum.LENDER)
 @Status(400).Description(OASDocs.STATUS_400_DESCR_VALIDATION)
 @Status(401).Description(OASDocs.STATUS_401_DESCR)
-export class TransactionsPortfolioCreditController extends BaseController {
+export class PortfolioTransactionsCreditController extends BaseController {
   constructor(@Inject() private service: CreditPortfolioTransactionService) {
     super();
   }
